+++ inst/run/schedulers/multicore.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-# Run the pipeline in a persistent background process:
+# Run this script in a shell with ./run.sh
+
+# This script runs the pipeline in a persistent background process:
 nohup nice -4 R CMD BATCH run.R &
 
 # Change the nice level above as appropriate